Answer player and backlight changes for profile

+++ answer_player.py
@@ -2,31 +2,36 @@
 # project: pRodriguezAssistant
 import subprocess
 import time
-from backlight_control import BacklightControl
 
 class AnswerPlayer:
     lang = 'en'
-    def __init__(self, lang, audio_files):
+    def __init__(self, lang, audio_files, mouth_bl, eyes_bl):
         self.mic_gain = 40
         self.mic_set(self.mic_gain)
-        self.mouth_bl = BacklightControl('MOUTH')
-        self.eyes_bl = BacklightControl('EYES')
-        self.mouth_bl.exec_cmd('OFF')
+        self.eyes_bl = eyes_bl
+        self.mouth_bl = mouth_bl
+        if self.mouth_bl:
+            self.mouth_bl.exec_cmd('OFF')
         self.audio_files = audio_files
         AnswerPlayer.lang = lang
 
     def play_wav(self, path, bl_command):
         aplay_exe = 'aplay -Dplug:default ' + str(path)
         aplay_proc = subprocess.Popen(["%s" % aplay_exe], shell=True, stdout=subprocess.PIPE)
-        mouth_bl_proc = self.mouth_bl.exec_cmd(bl_command)
 
         eyes_bl_proc = None
-        if bl_command == 'PLUGGED_IN':
-            eyes_bl_proc = self.eyes_bl.exec_cmd('BLINK_PLUGGED_IN')
+        mouth_bl_proc = None
+        if self.mouth_bl:
+            mouth_bl_proc = self.mouth_bl.exec_cmd(bl_command)
+        if self.eyes_bl:
+            if bl_command == 'PLUGGED_IN':
+                eyes_bl_proc = self.eyes_bl.exec_cmd('BLINK_PLUGGED_IN')
+
         code = aplay_proc.wait()
 
-        mouth_bl_proc.terminate()
-        if eyes_bl_proc != None:
+        if mouth_bl_proc:
+            mouth_bl_proc.terminate()
+        if eyes_bl_proc:
             eyes_bl_proc.terminate()
             self.eyes_bl.exec_cmd('ON')
 
@@ -48,7 +53,8 @@ class AnswerPlayer:
             self.mic_set(0)
             self.play_wav('./audio/' + AnswerPlayer.lang + '/' + answer + '.wav', bl_command)
             self.mic_set(self.mic_gain)
-            self.mouth_bl.exec_cmd('OFF')
+            if self.mouth_bl:
+                self.mouth_bl.exec_cmd('OFF')
         else:
             print('No answer to this question!')
